API-12392: existing tests moved around and fixed
