+++ spec/controllers/admin/consumers_spec.rb
@@ -2,10 +2,10 @@
 
 require 'rails_helper'
 
-RSpec.describe 'Admin::Api::V0::Consumers', type: :request do
+RSpec.describe 'Admin::ConsumersController', type: :request do
   describe 'Migrating Consumers' do
     it 'initializes migration of consumers into new structure' do
-      post '/platform-backend/admin/api/v0/consumers:migrate'
+      post '/platform-backend/admin/dashboard/consumers:migrate'
       expect(response.status).to eq(200)
     end
   end
@@ -15,7 +15,7 @@ RSpec.describe 'Admin::Api::V0::Consumers', type: :request do
       user = FactoryBot.create(:user)
       FactoryBot.create(:consumer, :with_apis, user_id: user.id)
 
-      post '/platform-backend/admin/api/v0/consumers/destroy_all'
+      post '/platform-backend/admin/dashboard/consumers/destroy_all'
       expect(response.status).to eq(302)
     end
   end
