+++ spec/controllers/admin/dashboard_spec.rb
