+++ spec/controllers/users/omniauth_callbacks_spec.rb
